+++ packages/008/src/008QWorker.js
@@ -1,6 +1,20 @@
 import PQueue from 'p-queue';
 
-import { tts } from './008Q';
+import { transcript } from '008Q';
+
+export const tts = async ({ audio }) => {
+  const remote = (await transcript({ wav: audio.remote })).map(item => ({
+    ...item,
+    channel: 'remote'
+  }));
+  const local = (await transcript({ wav: audio.local })).map(item => ({
+    ...item,
+    channel: 'local'
+  }));
+  const merged = [...remote, ...local].sort((a, b) => a.start - b.start);
+
+  return merged;
+};
 
 const QUEUE = new PQueue({ concurrency: 2 });
