+++ packages/008desktop/index.js
@@ -157,6 +157,7 @@ const createWindow = anchor => {
   mainWindow.setContentSize(0, 0);
   mainWindow.setMenuBarVisibility(false);
   mainWindow.loadURL(APP_URL);
+  mainWindow.webContents.openDevTools();
 
   mainWindow.on('close', ev => {
     if (!QUITTING) ev.preventDefault();
